[1/2] Tag AWS resources created with Terraform (#1552)
Amend Terraform configuration to tag all deployed resources that support
tagging.

Update documentation in README.md to specify that manually-created resources
should be tagged in a manner consistent with those maintained by Terraform.

If all tags were static, it would make sense to contain tags in the shared
Terraform configuration. Because the content of the tags depends on user config
it is reasonable to maintain helper code in `azul.deployment`.